refactor: 동아리 행사 도메인을 하위 패키지로 분리
