+++ src/main/java/in/koreatech/koin/domain/club/event/enums/ClubEventStatus.java
@@ -1,12 +1,12 @@
-package in.koreatech.koin.domain.club.enums;
+package in.koreatech.koin.domain.club.event.enums;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum ClubEventType {
-    RECENT("최신 등록순", 0),
+public enum ClubEventStatus {
+    SOON("곧 행사 진행", 1),
     ONGOING("행사 진행 중", 1),
     UPCOMING("행사 예정", 2),
     ENDED("종료된 행사", 3)
