+++ src/main/java/in/koreatech/koin/domain/club/event/repository/ClubEventImageRepository.java
@@ -1,11 +1,9 @@
-package in.koreatech.koin.domain.club.repository;
-
-import java.util.List;
+package in.koreatech.koin.domain.club.event.repository;
 
 import org.springframework.data.repository.Repository;
 
-import in.koreatech.koin.domain.club.model.ClubEvent;
-import in.koreatech.koin.domain.club.model.ClubEventImage;
+import in.koreatech.koin.domain.club.event.model.ClubEvent;
+import in.koreatech.koin.domain.club.event.model.ClubEventImage;
 
 public interface ClubEventImageRepository extends Repository<ClubEventImage, Integer> {
